+++ src/main/java/org/example/first/ExchangeService.java
@@ -1,5 +1,6 @@
 package org.example.first;
 
+import java.math.BigDecimal;
 import java.sql.SQLException;
 
 public class ExchangeService {
@@ -10,13 +11,13 @@ public class ExchangeService {
 
     public ExchangeDTO exchange(String baseCurrencyCode, String targetCurrencyCode, double amount) throws SQLException {
         if(exchangeRateDAO.isExists(baseCurrencyCode, targetCurrencyCode)) {
-            return exchangeDAO.getRate(baseCurrencyCode, targetCurrencyCode, amount);
+            return exchangeDAO.getRate(baseCurrencyCode, targetCurrencyCode, BigDecimal.valueOf(amount));
         }
         if(exchangeRateDAO.isExists(targetCurrencyCode, baseCurrencyCode)) {
-            return exchangeDAO.getRateFromReverseRate(targetCurrencyCode, baseCurrencyCode, amount);
+            return exchangeDAO.getRateFromReverseRate(targetCurrencyCode, baseCurrencyCode, BigDecimal.valueOf(amount));
         }
         if(exchangeRateDAO.isExists(INTERMEDIATE_CURRENCY_CODE, baseCurrencyCode) && exchangeRateDAO.isExists(INTERMEDIATE_CURRENCY_CODE, targetCurrencyCode)) {
-            return exchangeDAO.getRateWithIntermediate(baseCurrencyCode, targetCurrencyCode, INTERMEDIATE_CURRENCY_CODE, amount);
+            return exchangeDAO.getRateWithIntermediate(baseCurrencyCode, targetCurrencyCode, INTERMEDIATE_CURRENCY_CODE, BigDecimal.valueOf(amount));
         }
         throw new ElementNotFoundException();
     }
